refactor(entranceActions): refactor entrance actions
