+++ packages/web-app/src/actions/Entrance/ResetEntrance.js
@@ -1,0 +1,5 @@
+export const RESET_ENTRY_STATE = 'RESET_ENTRY_STATE';
+
+export const resetEntranceState = () => ({
+  type: RESET_ENTRY_STATE
+});
