+++ packages/web-app/src/components/appli/MoveEntranceToCave/index.jsx
@@ -7,7 +7,7 @@ import { useParams } from 'react-router-dom';
 import { isEmpty } from 'ramda';
 import { Divider } from '@material-ui/core';
 import Layout from '../../common/Layouts/Fixed/FixedContent';
-import { fetchEntry } from '../../../actions/Entry';
+import { fetchEntry } from '../../../actions/Entrance/GetEntrance';
 import MoveEntranceToCaveForm from './MoveEntranceToCaveForm';
 import Alert from '../../common/Alert';
 import AuthChecker from '../AuthChecker';
